Dummy for demultiplexing in place.
Right now demultiplexing will not actually run bcl2fastq, but it will
output the commandline used to stdout.

The somewhat hacky solution to setup authentication is to pass it along
in the workflow, I'd much rather figure out a way to do this using a
centralized config, but I don't know how to do that at the moment.