poll_status to use st2 style log and return json
I've changed the poll status script to use the st2 style of logging, and
returning status from the action as a tupple of (return_code, result) to
pass information. This way we can return relevant information, e.g.
stdout and stderr more easily.

I've also let the action return the results of the request as json also
on a successful result - to e.g. get the sisyphus check index results
back even when it is successful.