Changes preping to use the real mover
This introduces a option to the delivery workflows to skip calling mover
when running a delivery. This is so that we can do light integration
testing without running a real delivery.

It also separates the checking of delivery status from the starting a
delivery, so that workflows can be restarted from the status checking
stage if we experience intermitent failures.