New Arteria packs code
This is a transfer of all the code developer under github.com/johandahlberg/arteria-incubation

This removes all the old SNP&SEQ Technology Platform specific code which was in
this repo. It also introduces a new docker-compose based way to getting
started with Arteria which should hopefully make it easier to get
started for new people.